Refactor: move ConfettiParties to util package
This commit moves the `ConfettiParties.kt` file from the `feature/game/src/main/java/com/example/feature/game/components` package to `feature/game/src/main/java/com/example/feature/game/util`.

Additionally, the import for `ConfettiParties` in `VictoryDialog.kt` has been updated to reflect this change. One unused import for `Preview` was also removed from `PostGameActions.kt`.

This change improves the organization of utility classes within the `feature/game` module.
